refactor: app module (#83)
