+++ app/src/androidTest/java/jp/mydns/kokoichi0206/sakamichiapp/presentation/member_list/MemberListScreenTest.kt
@@ -10,7 +10,6 @@ import dagger.hilt.android.testing.HiltAndroidTest
 import dagger.hilt.android.testing.UninstallModules
 import jp.mydns.kokoichi0206.sakamichiapp.di.AppModule
 import jp.mydns.kokoichi0206.sakamichiapp.presentation.MainActivity
-import jp.mydns.kokoichi0206.sakamichiapp.presentation.util.Constants
 import jp.mydns.kokoichi0206.sakamichiapp.presentation.util.Screen
 import jp.mydns.kokoichi0206.sakamichiapp.presentation.util.getJsonFromMember
 import io.mockk.MockKAnnotations
