+++ app/src/main/java/jp/mydns/kokoichi0206/sakamichiapp/presentation/blog/BlogViewModel.kt
@@ -3,9 +3,9 @@ package jp.mydns.kokoichi0206.sakamichiapp.presentation.blog
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import dagger.hilt.android.lifecycle.HiltViewModel
+import jp.mydns.kokoichi0206.common.GroupName
 import jp.mydns.kokoichi0206.common.Resource
 import jp.mydns.kokoichi0206.domain.usecase.get_blogs.GetBlogsUseCase
-import jp.mydns.kokoichi0206.sakamichiapp.presentation.member_list.GroupName
 import kotlinx.coroutines.flow.*
 import javax.inject.Inject
