+++ app/src/main/java/jp/mydns/kokoichi0206/sakamichiapp/presentation/member_detail/MemberDetailScreen.kt
@@ -8,7 +8,7 @@ import androidx.compose.runtime.collectAsState
 import androidx.compose.runtime.getValue
 import androidx.compose.ui.Modifier
 import androidx.hilt.navigation.compose.hiltViewModel
-import jp.mydns.kokoichi0206.sakamichiapp.presentation.ui.theme.CustomSakaTheme
+import jp.mydns.kokoichi0206.common.ui.theme.CustomSakaTheme
 
 /**
  * Function to display member detail page.
