+++ app/src/main/java/jp/mydns/kokoichi0206/sakamichiapp/presentation/quiz/PlayQuizScreen.kt
@@ -27,12 +27,11 @@ import coil.compose.AsyncImage
 import coil.disk.DiskCache
 import coil.memory.MemoryCache
 import coil.request.ImageRequest
+import jp.mydns.kokoichi0206.common.GroupName
 import jp.mydns.kokoichi0206.sakamichiapp.R
-import jp.mydns.kokoichi0206.data.remote.LoggingInterceptor
-import jp.mydns.kokoichi0206.data.remote.RetryInterceptor
-import jp.mydns.kokoichi0206.sakamichiapp.presentation.ui.theme.SpaceMedium
-import jp.mydns.kokoichi0206.sakamichiapp.presentation.ui.theme.SpaceSmall
-import jp.mydns.kokoichi0206.sakamichiapp.presentation.ui.theme.Typography
+import jp.mydns.kokoichi0206.common.ui.theme.SpaceMedium
+import jp.mydns.kokoichi0206.common.ui.theme.SpaceSmall
+import jp.mydns.kokoichi0206.common.ui.theme.Typography
 import jp.mydns.kokoichi0206.sakamichiapp.presentation.util.TestTags
 import okhttp3.OkHttpClient
 import java.util.concurrent.TimeUnit
