+++ app/src/main/java/jp/mydns/kokoichi0206/sakamichiapp/presentation/setting/SettingsScreen.kt
@@ -13,10 +13,10 @@ import androidx.hilt.navigation.compose.hiltViewModel
 import com.google.accompanist.navigation.animation.composable
 import com.google.accompanist.navigation.animation.AnimatedNavHost
 import com.google.accompanist.navigation.animation.rememberAnimatedNavController
+import jp.mydns.kokoichi0206.common.Constants
 import jp.mydns.kokoichi0206.sakamichiapp.R
 import jp.mydns.kokoichi0206.sakamichiapp.presentation.setting.pages.*
-import jp.mydns.kokoichi0206.sakamichiapp.presentation.ui.theme.CustomSakaTheme
-import jp.mydns.kokoichi0206.sakamichiapp.presentation.util.Constants
+import jp.mydns.kokoichi0206.common.ui.theme.CustomSakaTheme
 
 @ExperimentalAnimationApi
 @Composable
