+++ app/src/main/java/jp/mydns/kokoichi0206/sakamichiapp/presentation/setting/pages/AboutAppScreen.kt
@@ -15,15 +15,14 @@ import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.unit.dp
 import androidx.navigation.NavHostController
-import jp.mydns.kokoichi0206.sakamichiapp.BuildConfig
+import jp.mydns.kokoichi0206.common.Constants
 import jp.mydns.kokoichi0206.sakamichiapp.R
 import jp.mydns.kokoichi0206.sakamichiapp.presentation.setting.SettingsUiState
 import jp.mydns.kokoichi0206.sakamichiapp.presentation.setting.components.SettingTopBar
-import jp.mydns.kokoichi0206.sakamichiapp.presentation.ui.theme.SpaceLarge
-import jp.mydns.kokoichi0206.sakamichiapp.presentation.ui.theme.SpaceMedium
-import jp.mydns.kokoichi0206.sakamichiapp.presentation.ui.theme.SpaceSmall
-import jp.mydns.kokoichi0206.sakamichiapp.presentation.ui.theme.TextGray
-import jp.mydns.kokoichi0206.sakamichiapp.presentation.util.Constants
+import jp.mydns.kokoichi0206.common.ui.theme.SpaceLarge
+import jp.mydns.kokoichi0206.common.ui.theme.SpaceMedium
+import jp.mydns.kokoichi0206.common.ui.theme.SpaceSmall
+import jp.mydns.kokoichi0206.common.ui.theme.TextGray
 import jp.mydns.kokoichi0206.sakamichiapp.presentation.util.TestTags
 
 /**
