+++ app/src/main/java/jp/mydns/kokoichi0206/sakamichiapp/presentation/util/TestTags.kt
@@ -62,7 +62,4 @@ object TestTags {
 
     // Setting utils
     const val SETTING_TOP_BAR = "SETTING_TOP_BAR"
-
-    // Utils
-    const val GROUP_BAR = "GROUP_BAR"
 }
