+++ core/common/build.gradle
@@ -6,8 +6,22 @@ plugins {
 android {
     namespace 'jp.mydns.kokoichi0206.common'
     compileSdk build_versions.target_sdk
+
+    defaultConfig {
+        minSdk build_versions.min_sdk
+    }
+    composeOptions {
+        kotlinCompilerExtensionVersion build_versions.compose
+    }
+    buildFeatures {
+        compose true
+    }
 }
 
 dependencies {
+    implementation deps.compose.ui
+    implementation deps.material.compose
+    implementation deps.material.material
+
     // modules
 }
