+++ core/common/src/main/java/jp/mydns/kokoichi0206/common/GroupName.kt
@@ -1,0 +1,7 @@
+package jp.mydns.kokoichi0206.common
+
+enum class GroupName(val jname: String) {
+    NOGIZAKA("乃木坂"),
+    SAKURAZAKA("櫻坂"),
+    HINATAZAKA("日向坂"),
+}
