+++ core/common/src/main/java/jp/mydns/kokoichi0206/common/TestTags.kt
@@ -1,0 +1,10 @@
+package jp.mydns.kokoichi0206.common
+
+/**
+ * TestTags for AndroidTest.
+ */
+object TestTags {
+
+    // Utils
+    const val GROUP_BAR = "GROUP_BAR"
+}
