+++ core/common/src/main/java/jp/mydns/kokoichi0206/common/ui/theme/Dimensions.kt
@@ -1,4 +1,4 @@
-package jp.mydns.kokoichi0206.sakamichiapp.presentation.ui.theme
+package jp.mydns.kokoichi0206.common.ui.theme
 
 import androidx.compose.ui.unit.dp
