+++ core/common/src/main/java/jp/mydns/kokoichi0206/common/ui/theme/Type.kt
@@ -1,4 +1,4 @@
-package jp.mydns.kokoichi0206.sakamichiapp.presentation.ui.theme
+package jp.mydns.kokoichi0206.common.ui.theme
 
 import androidx.compose.material.Typography
 import androidx.compose.ui.text.TextStyle
