+++ core/data/src/main/java/jp/mydns/kokoichi0206/data/remote/LoggingInterceptor.kt
@@ -17,11 +17,11 @@ class LoggingInterceptor : Interceptor {
     override fun intercept(chain: Interceptor.Chain): Response {
         val request: Request = chain.request()
         val t1 = System.nanoTime()
-        Timber.tag(tag).d("${request.url()}, ${request.url()}")
+        Timber.tag(tag).d("${request.url}, ${request.url}")
 
         val response: Response = chain.proceed(request)
         val t2 = System.nanoTime()
-        Timber.tag(tag).d("${response.request().url()}, ${(t2 - t1) / 1e6}, ${response.headers()}")
+        Timber.tag(tag).d("${response.request.url}, ${(t2 - t1) / 1e6}, ${response.headers}")
         Timber.tag(tag).d("${(t2 - t1) / 1e6} m sec")
 
         return response
