+++ core/model/build.gradle
@@ -11,4 +11,8 @@ android {
 
 dependencies {
     implementation deps.room.room
+    implementation deps.retrofit.gson
+
+    // modules
+    implementation project(":core:common")
 }
